pyproject.toml
[project]
name = "athm_tip"
version = "0.1.0"
description = "Example site demonstrating django-athm usage"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.8",
    "django-anymail[amazon-ses]>=13.1",
    "django-athm==1.0.0b1",
    "django-tailwind[cookiecutter,honcho,reload]>=4.4.1",
    "python-dotenv>=1.2.1",
    "sentry-sdk>=2.44.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "commitizen>=4.10.0",
    "pre-commit>=4.4.0",
    "ruff>=0.14.5",
]
